Refactor: move notifications logic into a seperate controller
