+++ app/controllers/support_interface/provider_users_controller.rb
@@ -71,29 +71,6 @@ module SupportInterface
       @provider_user = ProviderUser.find(params[:provider_user_id])
     end
 
-    def toggle_notifications
-      provider_user = ProviderUser.find(params[:provider_user_id])
-
-      SaveProviderUserNotificationPreferences
-        .new(provider_user: provider_user)
-        .backfill_notification_preferences!(send_notifications: !provider_user.send_notifications)
-
-      flash[:success] = 'Provider user updated'
-      redirect_to support_interface_provider_user_path(provider_user)
-    end
-
-    def update_notifications
-      render_404 unless FeatureFlag.active?(:configurable_provider_notifications)
-
-      provider_user = ProviderUser.find(params[:provider_user_id])
-      notification_preferences = provider_user.notification_preferences
-
-      if notification_preferences.update!(notification_preferences_params)
-        flash[:success] = 'Provider user notifications updated'
-        redirect_to support_interface_provider_user_path(provider_user)
-      end
-    end
-
     def impersonate
       @provider_user = ProviderUser.find(params[:provider_user_id])
       dfe_sign_in_user.begin_impersonation! session, @provider_user
@@ -132,13 +109,6 @@ module SupportInterface
             .permit(provider_permissions_forms: {})
             .fetch(:provider_permissions_forms, {})
             .to_h
-    end
-
-    def notification_preferences_params
-      return ActionController::Parameters.new unless params.key?(:provider_user_notification_preferences)
-
-      params.require(:provider_user_notification_preferences)
-        .permit(ProviderUserNotificationPreferences::NOTIFICATION_PREFERENCES)
     end
   end
 end
